Remove runtime.txt
A snafu occurred in the RStudio session that launced in mybinder.
First, despite being told that 4.1 was not a possibility, and despite
changing to 4.0 in runtime.txt, R version 4.1.1 still launched, but then
a new error occurred, and seemed to be related to RStudio being out of
date.

I wonder if removing the runtime.txt entirely will help.
Just an experiment.